% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, similar to \code{lm()}.}

\item{data}{Data frame input to be fitted.}
}
\value{
a table similar to the coefficent table from \code{summary()} with rows
for each coefficient and columns for the \code{Estimate}, \code{Std. Error},
\code{t value}, and \code{Pr(>|t|)}. There should be row and column names.
}
\description{
This function fits a linear model.
}
\examples{
my_lm(formula = mtcars$mpg ~ mtcars$cyl + mtcars$wt, data = mtcars)

}
\keyword{prediction}
