% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{My Random Forest Cross Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input representing the number of folds.}
}
\value{
Numeric indicating the cross validation MSE of the random forest models of fold \code{k}.
}
\description{
This function outputs the cross validation mean-squared error from random forest models.
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{inference.}
