% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{My t test}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept \code{"two.sided"}, \code{"less"}, or \code{"greater"}.
Otherwise, your function should throw an informative error.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list object containing:
\itemize{
\item \code{test_stat}: the numeric test statistic.
\item \code{df}: the degrees of freedom.
\item \code{alternative}: the value of the parameter \code{alternative}.
\item \code{p_val}: the numeric p-value.
}
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(t.test(x = mtcars$mpg, alternative = "two.sided",  mu = 20))
my_t.test(t.test(x = mtcars$mpg, alternative = "one.sided",  mu = 10))

}
\keyword{inference.}
