% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My k-Nearest Neighbors Cross Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame input of variables used to predict.}

\item{cl}{Data frame input of true variables being predicted.}

\item{k_nn}{Numeric input representing the number of neighbors.}

\item{k_cv}{Numeric input representing the number of folds.}
}
\value{
List object containing:
\itemize{
\item \code{class} a vector of the predicted class Y for all observations.
\item \code{cv_err} a numeric with the cross-validation misclassification error.
}
}
\description{
This function outputs a predicted class using given covariates and the cross
validation error associated with 1-knn and the given k-nn.
}
\examples{
penguins <- tidyr::drop_na(palmerpenguins::penguins)
data1 <- penguins[c("bill_length_mm", "bill_depth_mm", "flipper_length_mm", "body_mass_g")]
cl1 <- penguins["species"]
my_knn_cv(data1, cl1, 5, 5)

}
\keyword{inference,}
\keyword{prediction.}
